move out manage host filters into their own component
