+++ frontend/pages/hosts/ManageHostsPage/components/HostsFilterBlock/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HostsFilterBlock";
